feat:add clickMouse capsLock and delete mistake
