+++ src/index.js
@@ -4,6 +4,8 @@ import Keyboard from "./modules/virtual-keyboard"
 import MainContainer from "./modules/main-container"
 import TetArea from "./modules/textArea"
 import DescriptionText from "./modules/description"
+import {clickItem} from "./modules/item"
+import {itemToUpperCase} from "./modules/item"
 
 const mainContainer = new MainContainer()
 const textArea = new TetArea()
@@ -12,6 +14,7 @@ const keyboard = new Keyboard()
 
 const main = mainContainer.createElement()
 let clickMouse = false
+let capsLock = false
 
 document.body.append(main)
 main.append(textArea.createElement())
@@ -21,14 +24,22 @@ main.append(descriptionText.createElement())
 document.addEventListener('keydown',(e) => {
     e.preventDefault()
     clickMouse = false
+    capsLock = true
     let elem = document.getElementById(`${e.code}`)
-    clickItem(elem)
+    clickItem(elem,clickMouse,capsLock)
 })
 document.addEventListener('keyup',(e) => {
     e.preventDefault()
     clickMouse = false
+    capsLock = false
     let elem = document.getElementById(`${e.code}`)
-    elem.classList.remove('clickItem')
+    if(elem.dataset.item === 'CapsLock' & capsLock == false){
+        itemToUpperCase()
+        elem.classList.remove('bg-caps')
+        elem.classList.add('clickItem')
+    }else{
+        elem.classList.remove('clickItem')
+    }
 })
 document.addEventListener('click',(e) => {
     if(e.target.classList.contains('item')) {
@@ -36,60 +47,19 @@ document.addEventListener('click',(e) => {
         clickMouse = true
         const idElem = e.target.id
         let elem = document.getElementById(`${idElem}`)
+        if(elem.dataset.item === 'CapsLock' & !elem.classList.contains('bg-caps')) {
+            elem.classList.add('bg-caps')
+            itemToUpperCase()
+        }else if (elem.dataset.item === 'CapsLock' & elem.classList.contains('bg-caps')) {
+            console.log('good')
+            itemToUpperCase()
+            elem.classList.remove('bg-caps')
+            elem.classList.add('clickItem')
+        }
         clickItem(elem,clickMouse)
     }
 })
 
-function clickItem(event,clickMouse) {
-    const inputText = document.querySelector('#inputText')
-    const backSpace = document.getElementById('CapsLock')
-    const backSpaceUp =  backSpace.classList.contains('bg-caps')
-    const valueItem =  backSpaceUp ? event.textContent.toUpperCase() : event.textContent
-       
-    if((event.classList.contains('false')) !== true){
-        inputText.value += valueItem
-    }
-    if(event.dataset.item === 'Backspace') {
-        let arr= inputText.value.split('')
-        arr.pop()
-        inputText.value = arr.join('')   
-    }
-    if(event.dataset.item === 'Tab') {
-        let tab = ' ' + ' ' + ' ' + ' '
-        inputText.value += tab
-    }
-    if(event.dataset.item === 'Enter') {
-        let entr = '\n'
-        inputText.value +=entr
-    }
-    if(event.classList.contains('back-slash')){
-        inputText.value += valueItem
-    }
-    if(event.dataset.item === 'CapsLock') {
-        let lock = document.querySelectorAll('.lock')
-        for(let  i of lock) {
-            if(i.classList.contains('lock') & i.classList.contains('Up') === false) {
-                i.classList.add('Up')
-            } else {
-                i.classList.remove('Up')
-            }
-        }
-        if(event.classList.contains('bg-caps')) {
-            event.classList.remove('bg-caps')
-            }else{
-                event.classList.add('bg-caps')
-            }   
-        }
-        const item = event.classList.contains('item')
-        if(item) {
-            event.classList.add('clickItem')
-        }
-        if(clickMouse){
-            event.addEventListener('animationend', () => {
-            event.classList.remove('clickItem')
-            })
-        }
-        
-    } 
+
